+++ .github/workflows/hw-ci/run_hw_ci.sh
@@ -7,8 +7,4 @@ HW_SUITE=$1
 
 set -euo pipefail
 
-cd java_tools
-./gradlew :autotest:shadowJar
-cd ..
-
 java -cp java_console/autotest/build/libs/autotest-all.jar $HW_SUITE