+++ firmware/hw_layer/ports/stm32/stm32h7/cfg/mcuconf.h
@@ -127,7 +127,7 @@
 #define STM32_PLL3_FRACN_VALUE              0
 #define STM32_PLL3_DIVP_VALUE               2
 #define STM32_PLL3_DIVQ_VALUE               20
-#define STM32_PLL3_DIVR_VALUE               12
+#define STM32_PLL3_DIVR_VALUE               24
 
 /*
  * Core clocks dynamic settings (can be changed at runtime).
@@ -241,7 +241,7 @@
 /*
  * ADC driver system settings.
  */
-#define STM32_ADC_DUAL_MODE                 FALSE
+#define STM32_ADC_DUAL_MODE                 TRUE
 #define STM32_ADC_SAMPLES_SIZE              32
 #define STM32_ADC_USE_ADC12                 TRUE